Add multi-slope RIR analysis and use edr_loss
examples: add multislope-based decay estimation and RT60 comparison utilities (estimate_band_decay_times, shelving_group_rt60), import numpy and multislope, and integrate analysis into example_gfdn: normalize/pad target RIR, compute per-octave multi-slope RT60s, print band-wise comparison against the GFDN shelving groups, and change default target RIR path. Also replace MultiResoSTFT registration with edr_loss in the trainer. flamo/optimize/loss.py: ensure mel_stft is moved to the same dtype as input (to(self.device).to(x.dtype)) to avoid dtype mismatches.

+++ flamo/optimize/loss.py
@@ -645,7 +645,7 @@ class edr_loss(nn.Module):
             fmax=self.sample_rate // 2,
             n_mels=64,
             verbose=False,
-        ).to(self.device)
+        ).to(self.device).to(x.dtype)
 
         return mel_stft(x)
